Fixes edit form submition
